Move traits to their own submodules
